changed web3 helper to typescript
